requirements.txt
# Core dependencies - Production ready
pandas>=2.0.0
numpy>=1.24.0

# LLM and Embeddings - Free alternatives
langchain-community>=0.0.20
langchain-chroma>=0.1.0
langchain-text-splitters>=0.0.1
sentence-transformers>=2.2.0
chromadb>=0.4.0

# Text processing
transformers>=4.30.0
torch>=2.0.0

# Web interface
gradio>=4.0.0

# CLI utilities
tabulate>=0.9.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Development (optional)
jupyter>=1.0.0
ipywidgets>=8.0.0
notebook>=7.0.0
